Module3 wording edits (#38)
* Wording edits to Module 2

Edited the text in module 2 for clarity; added a link to the GH repo
for the Bootcamp, so readers can find the Dockerfile and source code
for the node.js application image.

* Merge remote-tracking branch 'kubernetes/gh-pages' into module2-wording-edits

* Merge remote-tracking branch 'kubernetes/gh-pages' into module3-wording-edits

* Merge remote-tracking branch 'kubernetes/gh-pages' into module3-wording-edits

* Update/improve wording in Module 3

Made significant changes to the order in which information is
introduced in Module 3, to avoid scattershot information dumping and
improve clarity.

Standardized on one tense and mode of address (present tense, direct
address to user in second-person).

* Merge remote-tracking branch 'kubernetes/gh-pages' into module3-wording-edits

* Addresses comments in #38

Addresses feedback from @pwittroc on Pull Request #38.